requirements.txt
# This file is @generated by PDM.
# Please do not edit it manually.

asgiref==3.7.2
attrs==23.1.0
Brotli==1.1.0
certifi==2023.11.17
charset-normalizer==3.3.2
cloudinary==1.37.0
django==5.0
django-cloudinary-storage==0.3.0
django-cors-headers==4.3.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.26.5
gunicorn==21.2.0
idna==3.6
inflection==0.5.1
jsonschema==4.20.0
jsonschema-specifications==2023.11.2
netifaces==0.11.0
packaging==23.2
Pillow==10.1.0
psycopg2-binary==2.9.9
pyjwt==2.8.0
python-dotenv==1.0.0
python-magic==0.4.27
pytz==2023.3.post1
PyYAML==6.0.1
referencing==0.32.0
requests==2.31.0
rpds-py==0.13.2
setuptools==69.0.2
six==1.16.0
sqlparse==0.4.4
typing-extensions==4.8.0
uritemplate==4.1.1
urllib3==2.1.0
whitenoise==6.6.0
